third_party/WebKit/LayoutTests/inspector/tracing/timeline-network/timeline-network-resource-details.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/timeline-test.js"></script>
<script>

function performActions()
{
    var image = new Image();
    var imagePromise = new Promise((fulfill) => image.onload = fulfill);
    image.src = "../resources/anImage.png";

    var script = document.createElement("script");
    script.src = "../../../http/tests/inspector/tracing/resources/timeline-network-resource.js";
    document.body.appendChild(script);
    var scriptPromise = new Promise((fulfill) => window.timelineNetworkResourceEvaluated = fulfill);

    return Promise.all([imagePromise, scriptPromise]);
}

function test()
{
    InspectorTest.invokeAsyncWithTimeline("performActions", finish);

    function finish()
    {
        var model = InspectorTest.timelineModel();
        var linkifier = new Components.Linkifier();
        function printRequestDetails(request)
        {
            return Timeline.TimelineUIUtils.buildNetworkRequestDetails(request, model, linkifier).then(printElement);
        }
        function printElement(element)
        {
            var rows = element.querySelectorAll(".timeline-details-view-row");
            for (var i = 0; i < rows.length; ++i) {
                var title = rows[i].firstChild.firstChild.textContent;
                var value = rows[i].lastChild.firstChild;
                if (value.children && value.children.length)
                  value = Array.from(value.children, element => element.textContent).slice(1).join('');
                else
                  value = value.textContent;
                if (title === "Duration" || title === "Mime Type")
                    value = typeof value;
                if (/^file:\/\//.test(value))
                    value = /[^/]*$/.exec(value)[0];
                InspectorTest.addResult(title + ": " + value);
            }
        }
        Promise.all(model.networkRequests().map(printRequestDetails)).then(InspectorTest.completeTest);
    }
}

</script>
</head>

<body onload="runTest()">
<p>
Tests the Timeline UI API for network requests.
</p>
</body>
</html>
