third_party/WebKit/LayoutTests/inspector/tracing/timeline-network/timeline-network-resource-details-expected.txt

Tests the Timeline UI API for network requests.

Script resource loaded
URL: timeline-network-resource.js
Duration: string
Request Method: GET
Priority: Low
Mime Type: string
Encoded Data: 223 B
Decoded Body: 223 B
Initiator: timeline-network-resource-details.html:15
URL: anImage.png
Duration: string
Request Method: GET
Priority: Low
Mime Type: string
Encoded Data: 12.9 KB
Decoded Body: 12.9 KB
Preview: 
